requirements.txt
Babel==2.10.1
certifi==2022.5.18
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
Flask==2.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-Table==0.5.0
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.4.0
pytz==2022.1
requests==2.27.1
SQLAlchemy==1.4.36
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
